+++ src/govuk/components/character-count/character-count.test.js
@@ -1,3 +1,5 @@
+import React from 'react';
+import { render } from '@testing-library/react';
 import examples from './examples';
 import { CharacterCount } from '.';
 import { diffComponentAgainstReferenceNunjucks } from '../../../../tests/utils/govuk-frontend-diff';
@@ -17,3 +19,14 @@ diffComponentAgainstReferenceNunjucks(
   CharacterCount,
   examples
 );
+
+describe('character count', () => {
+  it('correctly assigns a ref', () => {
+    const ref = React.createRef();
+    const { container } = render(
+      <CharacterCount {...examples.examples[0].data} ref={ref} />
+    );
+
+    expect(ref.current).toEqual(container.querySelector('textarea'));
+  });
+});