+++ src/govuk/components/radios/radios.test.js
@@ -16,25 +16,38 @@ examples.examples.forEach(example => {
 
 diffComponentAgainstReferenceNunjucks('radios', Radios, examples);
 
-it('Radios top level value prop maps correctly to checked items', () => {
-  const { getByDisplayValue } = render(
-    <Radios
-      value="no"
-      items={[
-        {
-          children: ['Yes'],
-          value: 'yes',
-        },
-        {
-          children: ['No'],
-          value: 'no',
-        },
-      ]}
-      name="yesno"
-      onChange={() => {}}
-    />
-  );
+describe('radios', () => {
+  it('top level value prop maps correctly to checked items', () => {
+    const { getByDisplayValue } = render(
+      <Radios
+        value="no"
+        items={[
+          {
+            children: ['Yes'],
+            value: 'yes',
+          },
+          {
+            children: ['No'],
+            value: 'no',
+          },
+        ]}
+        name="yesno"
+        onChange={() => {}}
+      />
+    );
 
-  expect(getByDisplayValue('yes').checked).toEqual(false);
-  expect(getByDisplayValue('no').checked).toEqual(true);
+    expect(getByDisplayValue('yes').checked).toEqual(false);
+    expect(getByDisplayValue('no').checked).toEqual(true);
+  });
+
+  it('correctly assigns a ref', () => {
+    const data = { ...examples.examples[0].data };
+    const ref = React.createRef();
+
+    data.items[0].ref = ref;
+
+    const { container } = render(<Radios {...data} />);
+
+    expect(ref.current).toEqual(container.querySelector('input'));
+  });
 });