+++ src/govuk/components/select/select.story.js
@@ -3,10 +3,11 @@ import { storiesOf } from '@storybook/react';
 import { Select as SelectComponent } from '.';
 import examples from './examples';
 import processExampleData from '../../../../utils/processExampleData';
+import { WithRef } from '../../../../utils/WithRef';
 
 const stories = storiesOf('select', module);
 
-const Select = props => {
+const Select = React.forwardRef((props, ref) => {
   const { value: initialValue, ...restProps } = props;
   const [value, setValue] = useState(initialValue);
 
@@ -15,12 +16,23 @@ const Select = props => {
   };
 
   return (
-    <SelectComponent onChange={onChangeHandler} value={value} {...restProps} />
+    <SelectComponent
+      ref={ref}
+      onChange={onChangeHandler}
+      value={value}
+      {...restProps}
+    />
   );
-};
+});
+
+Select.displayName = 'Select';
 
 for (const example of Object.values(
   processExampleData(examples.examples, 'select')
 )) {
   stories.add(example.name, () => <Select {...example.data} />);
 }
+
+stories.add('with ref', () => (
+  <WithRef Component={Select} {...examples.examples[0].data} />
+));